+++ dubbo-samples-mock/src/main/java/org/apache/dubbo/samples/mock/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.mock;
+package org.apache.dubbo.samples.mock;
 
 import java.io.File;
 import java.lang.reflect.Method;
