+++ dubbo-samples-mock/src/main/java/org/apache/dubbo/samples/mock/api/DemoService.java
@@ -15,19 +15,8 @@
  *   limitations under the License.
  */
 
-package com.alibaba.dubbo.samples.mock;
+package org.apache.dubbo.samples.mock.api;
 
-import com.alibaba.dubbo.samples.mock.api.DemoService;
-
-import org.springframework.context.support.ClassPathXmlApplicationContext;
-
-public class MockConsumer {
-
-    public static void main(String[] args) {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/mock-consumer.xml"});
-        context.start();
-        DemoService demoService = (DemoService) context.getBean("demoService"); // get remote service proxy
-        String hello = demoService.sayHello("world"); // call remote method
-        System.out.println(hello); // get result
-    }
+public interface DemoService {
+    String sayHello(String name);
 }
