+++ dubbo-samples-mock/src/main/resources/spring/mock-consumer.xml
@@ -31,6 +31,6 @@
 
     <!-- generate proxy for the remote service, then demoService can be used in the same way as the
     local regular interface -->
-    <dubbo:reference id="demoService" check="false" interface="com.alibaba.dubbo.samples.mock.api.DemoService" mock="true"/>
+    <dubbo:reference id="demoService" check="false" interface="org.apache.dubbo.samples.mock.api.DemoService" mock="true"/>
 
 </beans>
