+++ dubbo-samples-mock/src/main/resources/spring/mock-provider.xml
@@ -33,9 +33,9 @@
     <dubbo:protocol name="dubbo" port="20890"/>
 
     <!-- service implementation, as same as regular local bean -->
-    <bean id="demoService" class="com.alibaba.dubbo.samples.mock.impl.DemoServiceImpl"/>
+    <bean id="demoService" class="org.apache.dubbo.samples.mock.impl.DemoServiceImpl"/>
 
     <!-- declare the service interface to be exported -->
-    <dubbo:service interface="com.alibaba.dubbo.samples.mock.api.DemoService" ref="demoService"/>
+    <dubbo:service interface="org.apache.dubbo.samples.mock.api.DemoService" ref="demoService"/>
 
 </beans>
